Refactored profile page with offcanvas components and a useProfilePage hook
