+++ upnext-web-app/src/UpNext/types/invitation.d.ts
@@ -1,0 +1,11 @@
+import { Group } from "./group";
+
+export interface Invitation {
+  _id: string;
+  group: Group;
+  invitedBy: string;
+  invitedUser: string;
+  status: string; // "pending" | "accepted" | "rejected"
+  createdAt: string;
+  updatedAt: string;
+}
